Move prefix functions to util module
